Rename the phpunit class to PHP
