+++ src/Runner/PhptTestCase.php
@@ -19,6 +19,7 @@ use PHPUnit\Framework\SkippedTestError;
 use PHPUnit\Framework\SelfDescribing;
 use PHPUnit\Util\InvalidArgumentHelper;
 use PHPUnit\Util\PHP\AbstractPhpProcess;
+use PHPUnit\Util\PHP;
 use Throwable;
 
 /**
@@ -86,7 +87,7 @@ class PhptTestCase implements Test, SelfDescribing
         }
 
         $this->filename = $filename;
-        $this->phpUtil  = $phpUtil ?: AbstractPhpProcess::factory();
+        $this->phpUtil  = $phpUtil ?: new PHP();
     }
 
     /**
